ui/requirements.txt
streamlit==1.42.0
psycopg2-binary==2.9.9
pandas==2.2.3
plotly==5.24.1